Move the shared tooling config module under `workspace/tooling`
