+++ src/commands/init.ts
@@ -8,7 +8,7 @@ import * as Effect from "effect/Effect"
 import * as Predicate from "effect/Predicate"
 import * as Command from "effect/unstable/cli/Command"
 import * as Flag from "effect/unstable/cli/Flag"
-import type { ToolingConfigState } from "#lib/workspace/tooling-config.ts"
+import type { ToolingConfigState } from "#lib/workspace/tooling/config.ts"
 import github from "#lib/integrations/ci/github.ts"
 import vscode from "#lib/integrations/editors/vscode.ts"
 import zed from "#lib/integrations/editors/zed.ts"
