+++ src/lib/integrations/tooling/knip.ts
@@ -1,6 +1,6 @@
 import { getDependencyVersion } from "#lib/shared/version.macro.ts" with { type: "macro" }
 import { toKnipTsConfigContent } from "#lib/workspace/knip-config.ts"
-import { defineConfigTooling } from "#lib/workspace/tooling-config.ts"
+import { defineConfigTooling } from "#lib/workspace/tooling/config.ts"
 
 export default defineConfigTooling({
   configContent: () => toKnipTsConfigContent(),
